requirements.txt
django
fastapi
uvicorn
django-oauth-toolkit
pyjwt
djangorestframework-simplejwt[crypto]
# https://pypi.tuna.tsinghua.edu.cn/simple
